requirements.txt
ansible
redis
pyyaml
pylibmc
